// src/shared/api/generated/models/supportKnowledgeCitationDraftResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportKnowledgeCitationDraftResponseDtoMode } from "./supportKnowledgeCitationDraftResponseDtoMode";
import type { SupportKnowledgeCitationDraftResponseDtoState } from "./supportKnowledgeCitationDraftResponseDtoState";

export interface SupportKnowledgeCitationDraftResponseDto {
  actionEtag: string;
  documentId: string;
  expiresAt: string;
  id: string;
  mode: SupportKnowledgeCitationDraftResponseDtoMode;
  revisionId: string;
  /** @minimum 1 */
  revisionNumber: number;
  state: SupportKnowledgeCitationDraftResponseDtoState;
  /**
   * @maxLength 10000
   * @nullable
   */
  text?: string | null;
  /** @minimum 1 */
  version: number;
}